pyproject.toml
[project]
name = "farm-ai-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Sachin Tiwari",email = "mailtosachin1279@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (==22.6)",
    "python-dotenv (==1.2.1)",
    "requests (==2.32.5)",
    "pytz (==2025.2)",
    "faster-whisper (==1.2.1)",
    "google-genai (>=1.63.0,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
